generative.ai/developments/easy_llm_inference/requirements.txt
# LLM Standalone — Python dependencies
# Install:  pip install -r requirements.txt

torch>=2.0.0
transformers>=4.40.0
accelerate>=0.26.0
gradio>=4.0.0
llama-cpp-python>=0.3.0
